apps/api/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "facturia-api"
version = "2.0.0"
description = "FacturIA Invoice Processing API"
requires-python = ">=3.11"
authors = [
    {name = "FacturIA Team"}
]
readme = "README.md"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "sqlalchemy>=1.4.49",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "boto3>=1.34.34",
    "python-dotenv>=1.0.0",
    "structlog>=23.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "redis>=5.0.1",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "opencv-python>=4.9.0",
    "pillow>=10.2.0",
    "img2pdf>=0.5.1",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.21.0",
    "pdfplumber>=0.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.23.4",
    "pytest-cov>=4.1.0",
    "pytest-env>=1.1.3",
    "httpx>=0.26.0",
    "faker>=22.0.0",
    "factory-boy>=3.3.0",
    "moto>=4.2.14",
    "black>=24.1.1",
    "flake8>=7.0.0",
    "mypy>=1.8.0",
    "isort>=5.13.2",
]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
ignore_errors = true
